Moves the cookie header parsing method to a trait so that it can be shared Bumps version no. to 0.4.1 Adds throws tags to the relevant client methods
