+++ src/Action/SessionResolveRequest.php
@@ -3,11 +3,11 @@ namespace Citadel\Action;
 
 use Citadel\Exception\MalformedCookieHeaderException;
 use Citadel\Interface\CitadelRequestInstanceInterface;
+use Citadel\Trait\CookieHeaderParserTrait;
 
 class SessionResolveRequest implements CitadelRequestInstanceInterface
 {
-    private const string HOST_PATTERN = '/(?<host>__Host-CSID=[^;]+)/';
-    private const string SECURE_PATTERN = '/(?<secure>__Secure-CSIDa=[^;]+)/';
+    use CookieHeaderParserTrait;
 
     /**
      * @throws MalformedCookieHeaderException
@@ -18,29 +18,5 @@ class SessionResolveRequest implements CitadelRequestInstanceInterface
         public readonly string $clientSecret
     ) {
         $this->cookieHeader = $this->parseAndValidateCookieHeader($this->cookieHeader);
-    }
-
-    /**
-     * @throws MalformedCookieHeaderException
-     */
-    private function parseAndValidateCookieHeader(string $cookieHeader): string
-    {
-        $cookieHeaderComponentsFound = [];
-
-        if (preg_match(self::HOST_PATTERN, $cookieHeader, $hostMatches)) {
-            $cookieHeaderComponentsFound[] = $hostMatches['host'];
-        }
-
-        if (preg_match(self::SECURE_PATTERN, $cookieHeader, $secureMatches)) {
-            $cookieHeaderComponentsFound[] = $secureMatches['secure'];
-        }
-
-        if (empty($cookieHeaderComponentsFound)) {
-            throw new MalformedCookieHeaderException(
-                "The required components of the header value could not be found"
-            );
-        }
-
-        return implode('; ', $cookieHeaderComponentsFound) . ';';
     }
 }
