+++ src/Action/SessionRevokeRequest.php
@@ -1,13 +1,22 @@
 <?php
 namespace Citadel\Action;
 
+use Citadel\Exception\MalformedCookieHeaderException;
 use Citadel\Interface\CitadelRequestInstanceInterface;
+use Citadel\Trait\CookieHeaderParserTrait;
 
-readonly class SessionRevokeRequest implements CitadelRequestInstanceInterface
+class SessionRevokeRequest implements CitadelRequestInstanceInterface
 {
+    use CookieHeaderParserTrait;
+
+    /**
+     * @throws MalformedCookieHeaderException
+     */
     public function __construct(
         public string $cookieHeader,
-        public string $clientId,
-        public string $clientSecret
-    ) {}
+        public readonly string $clientId,
+        public readonly string $clientSecret
+    ) {
+        $this->cookieHeader = $this->parseAndValidateCookieHeader($this->cookieHeader);
+    }
 }
